добавил структуры и интерфейс
