+++ golang/lab/lab6.go
