+++ golang/lab/product.go
@@ -1,0 +1,8 @@
+package main
+
+type Product interface {
+	GetPrice() float64
+	SetPrice(price float64)
+	ApplyDiscount(percent float64)
+	GetDetails() string
+}
